move component under dir component
