+++ src/components/ResultPanel/ResultPanel.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactMarkdown from 'react-markdown';
 import gfm from "remark-gfm";
-import { ErrorTab } from "./ErrorTab";
+import { ErrorTab } from "../ErrorTab/ErrorTab";
 
 interface Props {
     content?: string;
